+++ src/hooks/template/index.ts
@@ -2,4 +2,8 @@ import { useAppMenu } from './useAppMenu'
 import { useMainPage } from './useMainPage'
 import { useMenuTag } from './useMenuTag'
 
+export type { MaximizeOptions } from './useMainPage'
+export type { Target } from './useAppMenu'
+export type { CloseMenuTag } from './useMenuTag'
+
 export { useAppMenu, useMainPage, useMenuTag }