+++ src/hooks/template/useMainPage.ts
@@ -17,6 +17,10 @@ import { useWindowSize } from '@vueuse/core'
 
 import type { Ref } from 'vue'
 
+export interface MaximizeOptions {
+  zIndex?: string
+}
+
 export function useMainPage() {
   /**
    *
@@ -45,19 +49,32 @@ export function useMainPage() {
    *
    * 该方法仅针对于 LayoutContent 区域，并且依赖全局属性 layoutContentMaximize
    */
-  const maximize = (full: boolean) => {
+  const maximize = (full: boolean, options?: MaximizeOptions) => {
     const contentEl = unrefElement(LAYOUT_CONTENT_REF as Ref<HTMLElement>)
 
     if (contentEl) {
       const { left, top } = contentEl.getBoundingClientRect() // 使用 left, top 计算 translate 偏移
       const { height } = useWindowSize() // 获取实际高度避免 100vh 会导致手机端浏览器获取不准确问题
+      const { zIndex = '99' } = options ?? {}
 
       full
         ? addStyle(contentEl, {
-            transform: `translate(-${left}px, -${top}px)`,
+            position: 'fixed',
+            width: '100%',
             height: `${height.value}px`,
+            transform: `translate(-${left}px, -${top}px)`,
+            transition: 'all 0.3s var(--r-bezier)',
+            zIndex,
           })
-        : removeStyle(contentEl, ['transform', 'height'])
+        : removeStyle(contentEl, [
+            'position',
+            'width',
+            'height',
+            'transform',
+            // 为了兼容浏览器 zIndex 的样式表
+            'zIndex',
+            'z-index',
+          ])
     }
 
     setVariable('layoutContentMaximize', full)