+++ src/utils/vue/watchEffectWithTarget.ts
@@ -22,8 +22,8 @@ import type { AnyFC } from '@/types/modules/utils'
  * 该方法使用 watchEffect 实现副作用函数的执行
  * 并且能够在 effect 作用域卸载时，自动停止监听
  */
-export function watchEffectWithTarget(
-  fc: AnyFC,
+export function watchEffectWithTarget<T extends AnyFC>(
+  fc: T,
   watchOptions?: WatchOptionsBase,
 ) {
   const stop = watchEffect(fc, watchOptions)